requirements.txt
streamlit
google-generativeai
requests
edge-tts
moviepy
pillow
numpy
trafilatura
feedparser
mutagen